Remove broken bit from CMakeLists.txt
Besides an obvious typo, this functionality was
never used because CMAKE_MINIMUM_REQUIRED(VERSION)
was too low.  Additionally, even if the typo was
fixed and the minimum cmake version were
raised, it would break in case OpenZen is linked
statically into a project with a different setting.

All in all it's better to remove this.